services/analytics/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11
fastapi==0.109.0
uvicorn==0.27.0
